refactor(ai): share Responses error normalization
